+++ R/RcppExports.R
@@ -9,10 +9,6 @@ rsqlite_disconnect <- function(con) {
     invisible(.Call('RSQLite_rsqlite_disconnect', PACKAGE = 'RSQLite', con))
 }
 
-rsqlite_get_exception <- function(con) {
-    .Call('RSQLite_rsqlite_get_exception', PACKAGE = 'RSQLite', con)
-}
-
 rsqlite_copy_database <- function(from, to) {
     invisible(.Call('RSQLite_rsqlite_copy_database', PACKAGE = 'RSQLite', from, to))
 }
@@ -39,6 +35,10 @@ rsqlite_fetch <- function(res, n = 10L) {
 
 rsqlite_bind_params <- function(res, params) {
     invisible(.Call('RSQLite_rsqlite_bind_params', PACKAGE = 'RSQLite', res, params))
+}
+
+rsqlite_find_params <- function(res, param_names) {
+    .Call('RSQLite_rsqlite_find_params', PACKAGE = 'RSQLite', res, param_names)
 }
 
 rsqlite_bind_rows <- function(res, params) {