+++ R/utils.R
@@ -1,3 +1,7 @@
 vlapply <- function(X, FUN, ..., USE.NAMES = TRUE) {
   vapply(X = X, FUN = FUN, FUN.VALUE = logical(1L), ..., USE.NAMES = USE.NAMES)
 }
+
+vcapply <- function(X, FUN, ..., USE.NAMES = TRUE) {
+  vapply(X = X, FUN = FUN, FUN.VALUE = character(1L), ..., USE.NAMES = USE.NAMES)
+}