+++ src/RcppExports.cpp
@@ -30,17 +30,6 @@ BEGIN_RCPP
     return R_NilValue;
 END_RCPP
 }
-// rsqlite_get_exception
-Rcpp::List rsqlite_get_exception(const XPtr<SqliteConnectionPtr>& con);
-RcppExport SEXP RSQLite_rsqlite_get_exception(SEXP conSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< const XPtr<SqliteConnectionPtr>& >::type con(conSEXP);
-    rcpp_result_gen = Rcpp::wrap(rsqlite_get_exception(con));
-    return rcpp_result_gen;
-END_RCPP
-}
 // rsqlite_copy_database
 void rsqlite_copy_database(const XPtr<SqliteConnectionPtr>& from, const XPtr<SqliteConnectionPtr>& to);
 RcppExport SEXP RSQLite_rsqlite_copy_database(SEXP fromSEXP, SEXP toSEXP) {
@@ -122,6 +111,18 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< List >::type params(paramsSEXP);
     rsqlite_bind_params(res, params);
     return R_NilValue;
+END_RCPP
+}
+// rsqlite_find_params
+IntegerVector rsqlite_find_params(const XPtr<SqliteResult>& res, CharacterVector param_names);
+RcppExport SEXP RSQLite_rsqlite_find_params(SEXP resSEXP, SEXP param_namesSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const XPtr<SqliteResult>& >::type res(resSEXP);
+    Rcpp::traits::input_parameter< CharacterVector >::type param_names(param_namesSEXP);
+    rcpp_result_gen = Rcpp::wrap(rsqlite_find_params(res, param_names));
+    return rcpp_result_gen;
 END_RCPP
 }
 // rsqlite_bind_rows