+++ src/SqliteResult.h
@@ -7,7 +7,6 @@
 #include "sqlite3/sqlite3.h"
 
 #include "SqliteConnection.h"
-#include "SqliteUtils.h"
 
 class SqliteResult : boost::noncopyable {
   sqlite3_stmt* pStatement_;
@@ -36,6 +35,7 @@ public:
 
 public:
   void bind(const Rcpp::List& params);
+  Rcpp::IntegerVector find_params(const Rcpp::CharacterVector& param_names);
   void bind_rows(const Rcpp::List& params);
   Rcpp::List fetch(int n_max = -1);
   Rcpp::List column_info();